add voucher

+++ src/main/java/storemanagement/example/group_15/app/dto/response/voucher/VoucherResponseDTO.java
@@ -8,5 +8,5 @@ import lombok.Data;
 public class VoucherResponseDTO {
   private Long id;
   private String name;
-  private Integer discountValue;
+  private Long discountValue;
 }
